+++ src/routes/Router/RouterPath.tsx
@@ -1,0 +1,13 @@
+export enum RouterPath {
+  OfferedApiDelegations = 'offered-api-delegations',
+  Overview = 'overview',
+  ChooseApi = 'choose-api',
+  ChooseOrg = 'choose-org',
+  Confirmation = 'confirmation',
+  Receipt = 'receipt',
+  ReceivedApiDelegations = 'received-api-delegations',
+  Profile = 'Profile',
+  BasePath = '/accessmanagement/ui',
+  ProfileAT23 = 'https://at23.altinn.cloud/ui/Profile',
+  ProfileProd = 'https://www.altinn.no/ui/Profile',
+}
